+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import { ChakraProvider } from "@chakra-ui/react";
 import { AppProps } from "next/app";
-import { theme } from "@masonmcelvain/ui/theme";
+import { theme } from "@ui/theme";
 
 function MyApp({ Component, pageProps }: AppProps) {
    return (
